pyproject.toml
[project]
name = "debatebet"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "instructor>=1.7.2",
    "openai>=1.62.0",
    "pydantic-settings>=2.7.1",
    "python-dotenv>=1.0.1",
    "pyyaml>=6.0.2",
    "tenacity>=9.0.0",
    "types-pyyaml>=6.0.12.20241230",
    "types-requests>=2.32.0.20241016",
]
